TypeResolver caches only open generic types instead of numerous combinations of closed types and no longer matches property types (see below). This improves the performance and also renders `_validateIncludingPropertyInfos` flag obsolete (removed). Issue #15
The following tests are failing:
- Resolved_types_should_be_equal_to_original_types
- Resolved_types_should_not_be_equal
- Resolved_type_A_and_B_should_be_different

TBD: it sounds impractical to have *non-anonymous* types with identical names and property names, which differ only in property types. Can we drop support of such cases and remove these remaining tests?
FYI: issue #6 was mainly addressing a problem of *anonymous* types for which it should still be sufficient to only match property names (without types, because types are always generic).
DECISION: remove irrelevant tests in the next commit!